Add 1 to internal TT
This means the sync pulses come _with_ timestamps of 0, 50ms, 100ms.
So the design registers `sync_time` as a whole multiple of the
sync period, rather than one less than this (which breaks the packetizer).
Whether this is strictly the _right_ thing to do isn't entirely obvious to me,
but the different is "just" a fixed offset in the timing of the system of 1 FPGA clock (4ns)