Bugfix 5
Only allow syncs to propagate on TelescopeTimes which are multiples
of 2^14. Packetizer requires at least 2^13 (number of FPGA clocks
which it takes to generate 1 packet's worth of samples).

Add MRST logic to reorder block (which has weird (large, prime)
internal reorder-orders) so that this block may only sync on the first
sync pulse after a MRST