v13.7
Increase LO precision to Fix45_44
Allow periodic loading.

The intention here is that we can now match the VLA's fshift
settings, which appear to be multiples of 100Hz, to a precision
of better than 1 mHz.
The new firmware allows the LO accumulators to be periodically
reset -- resetting every 1s would accumulate less than ~0.03 degrees
of phase error vs the VLA over the coarse of integration.